+++ superset/templates/superset/explorev2.html
@@ -9,7 +9,7 @@
 {% endblock %}
 
 {% block body %}
-  <link rel="stylesheet" type="text/css" href="/static/assets/stylesheets/exploreV2/explorev2.css">
+  <link rel="stylesheet" type="text/css" href="/static/assets/stylesheets/exploreV2/exploreV2.css">
   <div
     id="js-explore-view-container"
     data-bootstrap="{{ bootstrap_data }}"
